requirements.txt
datetime
google-generativeai
imbalanced-learn==0.12.2
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.2
plotly==5.20.0
pyarrow
python-dotenv
requests==2.32.3
scikit-learn==1.4.0
streamlit_lottie==0.0.5
streamlit==1.37.1
streamlit-navigation-bar==3.3.0
supabase==2.10.0
typing
xgboost
streamlit_shap
shap
matplotlib